Movie details view

+++ MovieTheatre/Controllers/RatingController.cs
@@ -18,15 +18,29 @@ namespace MovieTheatre.Controllers
         // GET: Rating
         public ActionResult Index(int userId = 0, int movieId = 0, int stars = 0)
         {
-            var ratings = db.Ratings.Where(s => s.UserID  == userId  || userId  == 0)
+            var isCurrentUserManager = (Boolean)Session["isCurrentUserManager"];
+
+            if (isCurrentUserManager == false)
+            {
+                return RedirectToAction("Index", "Error", new { message = "You're not allowed here!" });
+            }
+
+            var ratings = db.Ratings.Where(s => s.UserID == userId || userId == 0)
                                    .Where(s => s.MovieID == movieId || movieId == 0)
-                                   .Where(s => s.Stars   == stars   || stars   == 0);
+                                   .Where(s => s.Stars == stars || stars == 0);
             return View(ratings.ToList());
         }
 
         // GET: Rating/Details/5
         public ActionResult Details(int? id)
         {
+            var isCurrentUserManager = (Boolean)Session["isCurrentUserManager"];
+
+            if (isCurrentUserManager == false)
+            {
+                return RedirectToAction("Index", "Error", new { message = "You're not allowed here!" });
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -52,11 +66,18 @@ namespace MovieTheatre.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID,UserID,MovieID,Review,Stars,ReviewDate")] Rating rating)
         {
-            if (ModelState.IsValid)
+            var currentUserID = (int)Session["CurrentUserID"];
+            rating.ReviewDate = DateTime.Now;
+            rating.UserID = currentUserID;
+
+            if (currentUserID != 0)
             {
-                db.Ratings.Add(rating);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Ratings.Add(rating);
+                    db.SaveChanges();
+                    return Redirect("../User/Details/" + currentUserID);
+                }
             }
 
             return View(rating);
@@ -65,6 +86,13 @@ namespace MovieTheatre.Controllers
         // GET: Rating/Edit/5
         public ActionResult Edit(int? id)
         {
+            var isCurrentUserManager = (Boolean)Session["isCurrentUserManager"];
+
+            if (isCurrentUserManager == false)
+            {
+                return RedirectToAction("Index", "Error", new { message = "You're not allowed here!" });
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -84,6 +112,13 @@ namespace MovieTheatre.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ID,UserID,MovieID,Review,Stars,ReviewDate")] Rating rating)
         {
+            var isCurrentUserManager = (Boolean)Session["isCurrentUserManager"];
+
+            if (isCurrentUserManager == false)
+            {
+                return RedirectToAction("Index", "Error", new { message = "You're not allowed here!" });
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(rating).State = EntityState.Modified;
@@ -96,6 +131,13 @@ namespace MovieTheatre.Controllers
         // GET: Rating/Delete/5
         public ActionResult Delete(int? id)
         {
+            var isCurrentUserManager = (Boolean)Session["isCurrentUserManager"];
+
+            if (isCurrentUserManager == false)
+            {
+                return RedirectToAction("Index", "Error", new { message = "You're not allowed here!" });
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -113,6 +155,13 @@ namespace MovieTheatre.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            var isCurrentUserManager = (Boolean)Session["isCurrentUserManager"];
+
+            if (isCurrentUserManager == false)
+            {
+                return RedirectToAction("Index", "Error", new { message = "You're not allowed here!" });
+            }
+
             Rating rating = db.Ratings.Find(id);
             db.Ratings.Remove(rating);
             db.SaveChanges();
